python/requirements.txt
mysql-connector-python
pika
python-dotenv
fastapi
uvicorn
asyncio-mqtt
